Criar arquivo Metadata e inclusão da imagem para preview
Criar arquivo metadata.ts e importar para layout.ts
